User.service2

+++ CustomEd.User.Service/src/Controllers/StudentController.cs
@@ -1,10 +1,13 @@
 using Microsoft.AspNetCore.Mvc;
-using CustomEd.User.Service.Model;
-using CustomEd.User.Service.Data.Interfaces;
-using CustomEd.User.Service.Response;
 using CustomEd.User.Service.DTOs;
 using AutoMapper;
 using CustomEd.User.Service.Validators;
+using CustomEd.User.Service.PasswordService.Interfaces;
+using CustomEd.Shared.JWT.Interfaces;
+using Microsoft.AspNetCore.Authorization;
+using CustomEd.Shared.JWT;
+using CustomEd.Shared.Data.Interfaces;
+using CustomEd.Shared.Response;
 
 namespace CustomEd.User.Service.Controllers
 {
@@ -12,9 +15,10 @@ namespace CustomEd.User.Service.Controllers
     [Route("api/user/student")]
     public class StudentController : UserController<Model.Student>
     {
-        public StudentController(IGenericRepository<Model.Student> userRepository, IMapper mapper) : base(userRepository, mapper)
+        public StudentController(IGenericRepository<Model.Student> userRepository, IMapper mapper, IPasswordHasher passwordHasher, IJwtService jwtService) : base(userRepository, mapper, passwordHasher, jwtService)
         {
         }
+
 
         [HttpGet("student-id")]
         public async Task<ActionResult<SharedResponse<StudentDto>>> SearchStudentBySchoolId([FromQuery] string id)
@@ -36,12 +40,15 @@ namespace CustomEd.User.Service.Controllers
         public async Task<ActionResult<SharedResponse<Model.Student>>> CreateUser([FromBody] CreateStudentDto studentDto)
         {
 
-            var createStudentDtoValidator = new CreateStudentDtoValidator();
+            var createStudentDtoValidator = new CreateStudentDtoValidator(_userRepository);
             var validationResult = await createStudentDtoValidator.ValidateAsync(studentDto);
             if (!validationResult.IsValid)
             {
                 return BadRequest(SharedResponse<Model.Student>.Fail("Invalid input", validationResult.Errors.Select(e => e.ErrorMessage).ToList()));
             }
+            var passwordHash = _passwordHasher.HashPassword(studentDto.Password);
+            studentDto.Password = passwordHash;
+
 
             var student = _mapper.Map<Model.Student>(studentDto);
             
@@ -50,6 +57,8 @@ namespace CustomEd.User.Service.Controllers
             
         }
 
+
+        [Authorize]
         [HttpDelete("{id}")]
         public async Task<ActionResult<SharedResponse<StudentDto>>> RemoveUser(Guid id)
         {
@@ -58,11 +67,20 @@ namespace CustomEd.User.Service.Controllers
                 return BadRequest(SharedResponse<StudentDto>.Fail("Invalid Id", new List<string> { "Invalid id" }));
             }
 
+            var identityProvider = new IdentityProvider(HttpContext, _jwtService);
+            var currentUserId = identityProvider.GetUserId();
+
+            if(currentUserId != id)
+            {
+                return Unauthorized(SharedResponse<StudentDto>.Fail("Unauthorized", new List<string> { "Unauthorized" }));
+            }
+
 
             await _userRepository.RemoveAsync(id);
             return Ok(SharedResponse<StudentDto>.Success(null, "User deleted successfully"));
         }
 
+        [Authorize]
         [HttpPut]
         public async Task<ActionResult<SharedResponse<StudentDto>>> UpdateUser([FromBody] UpdateStudentDto studentDto)
         {
@@ -72,6 +90,18 @@ namespace CustomEd.User.Service.Controllers
             {
                 return BadRequest(SharedResponse<StudentDto>.Fail("Invalid input", validationResult.Errors.Select(e => e.ErrorMessage).ToList()));
             }
+
+            var identityProvider = new IdentityProvider(HttpContext, _jwtService);
+            var currentUserId = identityProvider.GetUserId();
+
+            if(currentUserId != studentDto.Id)
+            {
+                return Unauthorized(SharedResponse<StudentDto>.Fail("Unauthorized", new List<string> { "Unauthorized" }));
+            }
+
+            var passwordHash = _passwordHasher.HashPassword(studentDto.Password);
+            studentDto.Password = passwordHash;
+            studentDto.JoinDate = DateTime.UtcNow;
 
             var student = _mapper.Map<Model.Student>(studentDto);
             await _userRepository.UpdateAsync(student);
